refactor: wydziel useYearlyData z YearlySummary.jsx
Nowy hook src/hooks/useYearlyData.js (164 linii) hermetyzuje:
- fetch yearData + prevYearData + availableYears + compareData
- wszystkie memoizowane kalkulacje: miesięczne/kategoryjne agregacje,
  totals, YoY, cumulativeSavings, compareChartData, categoryCompareTable,
  personPieData
- cancellation na flagach dla obu effectów

YearlySummary.jsx: 636 → 543 linii (−93). Komponent jest teraz
sfokusowany na renderowaniu — całość logiki danych i kalkulacji w
testowalnym hooku.

Testy: 205/205. Build OK.
